Move shell handling functions to separate package
